Move API version validation into separate package
This commit moves version validation out of the v1beta1 package into
its own package. This code isn't part of the v1beta1 API, and contains
utilities that will be needed for the v1 API as well. No functional changes.
