+++ pkg/apis/version/version_validation.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package v1beta1
+package version
 
 import (
 	"context"
